Fucking encodings (and I guess on fucking Linux)
Close #17

On the server (Linux, I think Debian) the result of a `Redis.new.get` inside a
`Redis.new.multi` was returning a US-ASCII encoded string instead of a UTF-8
string. That caused all kinds of troubles (like exceptions when trying to print
the string) but mainly the ones described in #17.

The temporary fix forces the encoding to UTF-8 until a better solution is found.